+++ server/src/test/java/org/opensearch/indices/replication/SegmentReplicationSourceHandlerTests.java
@@ -18,6 +18,7 @@ import org.opensearch.action.ActionListener;
 import org.opensearch.cluster.metadata.IndexMetadata;
 import org.opensearch.cluster.node.DiscoveryNode;
 import org.opensearch.common.settings.Settings;
+import org.opensearch.common.util.CancellableThreads;
 import org.opensearch.index.shard.IndexShard;
 import org.opensearch.index.shard.IndexShardTestCase;
 import org.opensearch.index.store.StoreFileMetadata;
@@ -28,6 +29,8 @@ import org.opensearch.indices.replication.common.CopyState;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 
 import static org.mockito.Mockito.mock;
 
@@ -196,5 +199,48 @@ public class SegmentReplicationSourceHandlerTests extends IndexShardTestCase {
 
         handler.sendFiles(getSegmentFilesRequest, mock(ActionListener.class));
         Assert.assertThrows(OpenSearchException.class, () -> { handler.sendFiles(getSegmentFilesRequest, mock(ActionListener.class)); });
+    }
+
+    public void testCancelReplication() throws IOException, InterruptedException {
+        CountDownLatch latch = new CountDownLatch(1);
+        chunkWriter = mock(FileChunkWriter.class);
+
+        final ReplicationCheckpoint latestReplicationCheckpoint = primary.getLatestReplicationCheckpoint();
+        final CopyState copyState = new CopyState(latestReplicationCheckpoint, primary);
+        SegmentReplicationSourceHandler handler = new SegmentReplicationSourceHandler(
+            localNode,
+            chunkWriter,
+            threadPool,
+            copyState,
+            primary.routingEntry().allocationId().getId(),
+            5000,
+            1
+        );
+
+        final GetSegmentFilesRequest getSegmentFilesRequest = new GetSegmentFilesRequest(
+            1L,
+            replica.routingEntry().allocationId().getId(),
+            replicaDiscoveryNode,
+            Collections.emptyList(),
+            latestReplicationCheckpoint
+        );
+
+        // cancel before xfer starts. Cancels during copy will be tested in SegmentFileTransferHandlerTests, that uses the same
+        // cancellableThreads.
+        handler.cancel("test");
+        handler.sendFiles(getSegmentFilesRequest, new ActionListener<>() {
+            @Override
+            public void onResponse(GetSegmentFilesResponse getSegmentFilesResponse) {
+                Assert.fail("Expected failure.");
+            }
+
+            @Override
+            public void onFailure(Exception e) {
+                assertEquals(CancellableThreads.ExecutionCancelledException.class, e.getClass());
+                latch.countDown();
+            }
+        });
+        latch.await(2, TimeUnit.SECONDS);
+        assertEquals("listener should have resolved with failure", 0, latch.getCount());
     }
 }